+++ tests/test_utils/context.py
@@ -1,5 +1,13 @@
 import os
 import sys
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
 
-import rapidenv
+try:
+    # rapidenv installed
+    import rapidenv
+    __INSTALLED__ = True
+except:
+    # running from tests folder
+    sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
+    import rapidenv
+    __INSTALLED__ = False
+